refactor(saves): push helpers from SaveService into sub-services
Apply the LibraryService decomposition pattern (#432) to the SaveService
vertical. SaveService becomes a thin aggregate root that owns the live
SaveSyncState and composes the sub-services; every public callable
delegates. Drops the cyclic _save_service back-ref from every
sub-service.

Helper redistribution:
- Per-ROM save path resolution + local save discovery
  (_get_rom_save_info, _find_save_files, _resolve_retroarch_corename,
  _pending_sort_settings, _is_save_sort_changed) → new
  RomInfoService (services/saves/rom_info.py).
- Settings read/write (_is_save_sync_enabled, _get_server_device_id,
  get_save_sync_settings, update_save_sync_settings,
  _sanitize_setting) → StateService.
- Status + core-change reporting (get_save_status,
  check_save_status_background, check_core_change) → StatusService.
- Sync orchestration callables (pre_launch_sync, post_exit_sync,
  sync_rom_saves, sync_all_saves, resolve_sync_conflict,
  _resolve_conflict_*, _refresh_save_sort_state) → SyncEngine.
  SyncEngine now owns the _rom_sync_locks dict and the device
  registration flow (ensure_device_registered, list_devices) too,
  since every sync path falls back to registration when device_id
  is missing.
- Bulk local-save deletion stays on SaveService — it spans
  RomInfoService, the SaveFileAdapter, and StateService without
  belonging to any single sub-service.

Sub-service peer injection: SlotsService → StatusService (for
get_save_status / check_save_status_background); VersionsService →
SyncEngine (for _rom_lock dispatch). No new Protocols needed.

SaveService class: 1068 → 344 LOC. The public callable surface
exposed to main.py is unchanged.

Test reorganisation:
- TestFindSaveFiles + TestGetRomSaveInfo classes move from
  test_service.py to a new test_rom_info.py (21 tests, same coverage).
- _sanitize_setting tests in test_service.py now exercise the
  StateService location.
- test_bootstrap.py + test_plugin_saves.py + test_sync_engine.py
  updated to read sub-service-owned attrs through their new
  locations (_sync_engine._detect_sort_change,
  _rom_info._get_core_name, _state_svc.is_save_sync_enabled, etc.).
- test_plugin_saves._set_event_loop fixture now rebinds the
  sub-services' _loop too, since orchestration moved off the
  aggregate root.

All gates green: 2261 tests pass, ruff clean, basedpyright clean,
import-linter 7/7 contracts kept, cosmic call bans clean.
